requirements.txt
Django==3.0.5
zeep==3.4.0
#django rest :
djangorestframework==3.11.0
markdown==3.2.1
django-filter==2.2.0



